fixed bug with stack_push16 only pushing 1 byte of data instead of 2

+++ include/stack.h
@@ -3,7 +3,7 @@
 #include <stdint.h>
 
 void stack_push(uint8_t data);
-void stack_push16(uint8_t data);
+void stack_push16(uint16_t data);
 
 uint8_t stack_pop();
 uint16_t stack_pop16();
